Create weekdays program using switch
This program demonstrates the use of the switch statement in C.
The user can select a number corresponding to a day of the week
(1 for Monday, 2 for Tuesday, etc.), and the program prints
the name of the day. If the number is not between 1 and 7, it
displays a default message indicating an invalid choice.